Creates environment variables for server
Adds a step to the deployment workflow to create a `.env` file
with environment variables required by the server application.
This allows the server to access sensitive information such as API keys
and database credentials during runtime in the Heroku environment.
Also, it updates the JWT secret key variable name.